Reorganize the headers and include by new flavor. The header include each others from relative path, and others include from absolute path (specified by -I<dir> to compiler).
