+++ src/at_table.c
@@ -4,8 +4,8 @@
  *  \email tcath2s@gmail.com
  * */
 
-#include "../inc/at_command.h"
-#include "../inc/at_table.h"
+#include <at/at_command.h>
+#include <at/at_table.h>
 
 void at_table_register(at_cmd_context_t * context, at_cmd_cb_t table[], 
                 size_t count){
